freeを追加
