+++ .gitignore
@@ -1,3 +1,7 @@
 a.out
 .vscode
 minishell
+test
+out
+outfile
+infile
