+++ srcs/tokenize/tokenize.c
@@ -12,52 +12,20 @@
 
 #include "../../includes/minishell.h"
 
-t_token	*lasttoken(t_token **head)
+size_t	get_end_index_helper(char *line, size_t i)
 {
-	t_token	*token;
-
-	token = *head;
-	while (token)
+	if (line[i] == '|')
 	{
-		if (!token->next)
-			break ;
-		token = token->next;
+		while (line[i] == '|' && line[i] != '\0')
+			i++;
 	}
-	return (token);
-}
-
-void	addback(t_token **head, t_token *newtoken)
-{
-	t_token	*last_token;
-
-	last_token = lasttoken(head);
-	if (last_token)
-		last_token->next = newtoken;
 	else
-		*head = newtoken;
-}
-
-int	newtoken(t_token **head, char *line, size_t start, size_t end)
-{
-	t_token	*token;
-
-	if (start == end && line[end] == '\0')
-		return (1);
-	token = malloc(sizeof(t_token));
-	if (!token)
-		malloc_failed("malloc");
-	token->next = NULL;
-	token->is_read = false;
-	token->expanded_str = NULL;
-	token->str = ft_substr(line, start, (end - start));
-	if (!token->str)
-		malloc_failed("malloc");
-	if ((!ft_strncmp(token->str, "\'\'", 2) || !ft_strncmp(token->str, "\"\"", 2))
-		&& ft_strlen(token->str) == 2)
-		return (free_token(token));
-	set_token_kind(head, token);
-	addback(head, token);
-	return (0);
+	{
+		while (!is_space(line[i]) && !is_quotation_mark(line[i])
+			&& line[i] != '|' && line[i] != '\0')
+			i++;
+	}
+	return (i);
 }
 
 size_t	get_end_index(char *line, size_t i)
@@ -78,12 +46,13 @@ size_t	get_end_index(char *line, size_t i)
 		while (line[i] == c && line[i] != '\0')
 			i++;
 	}
-	else
+	else if (line[i] == '|')
 	{
-		while (!is_space(line[i]) && !is_quotation_mark(line[i])
-			&& line[i] != '\0')
+		while (line[i] == '|' && line[i] != '\0')
 			i++;
 	}
+	else
+		i = get_end_index_helper(line, i);
 	return (i);
 }
